+++ src/pages/SignUp/index.jsx
@@ -1,83 +1,35 @@
-import { useEffect, useRef } from 'react';
-
+import useSignUpPage from './hook';
 import Input from 'components/Common/Input/Input';
 import Title from 'components/Common/Title/Title';
 import Button from 'components/Common/Button/Button';
 import Fieldset from 'components/Common/Fieldset/Fieldset';
 import ValidateText from 'components/Common/ValidateText/ValidateText';
 import Form from 'components/Common/Form/Form';
-import useInputValidate from 'hooks/useInputValidate';
-import { useNavigate } from 'react-router-dom';
-import { PATH_NAME } from 'constants';
-import useAuth from 'hooks/useAuth';
-import useSnackBar from 'hooks/useSnackBar';
+
 import * as Styled from './style';
 
 const SignUp = () => {
-  const navigate = useNavigate();
-  const pwd = useRef(null);
-
-  const { showSuccessSnackBar, showErrorSnackBar } = useSnackBar();
-
-  const { isSignUpSucceed, isSignUpError, signUp, checkIsAuthenticated } =
-    useAuth();
-  const [emailValidate, handleEmailBlur] = useInputValidate('email');
-  const [nameValidate, handleNameBlur] = useInputValidate('name');
-  const [passwordValidate, handlePasswordBlur] = useInputValidate('password');
-  const [passwordCheckValidate, handlePasswordCheckBlur] =
-    useInputValidate('passwordCheck');
-
-  const handleSubmit = (e) => {
-    e.preventDefault();
-
-    const isAllValid =
-      emailValidate.isValid &&
-      nameValidate.isValid &&
-      passwordValidate.isValid &&
-      passwordCheckValidate.isValid;
-
-    if (!isAllValid) {
-      showErrorSnackBar('정보를 올바르게 입력하세요.');
-      return;
-    }
-
-    const {
-      email: { value: email },
-      name: { value: name },
-      password: { value: password },
-    } = e.target.elements;
-
-    signUp({
-      email,
-      name,
-      password,
-    });
-  };
-
-  useEffect(() => {
-    checkIsAuthenticated();
-  }, []);
-
-  useEffect(() => {
-    if (isSignUpSucceed) {
-      navigate(PATH_NAME.LOGIN);
-      showSuccessSnackBar('회원가입 성공');
-      return;
-    }
-    if (isSignUpError) {
-      showErrorSnackBar('입력한 정보를 확인 하세요.');
-      return;
-    }
-  }, [isSignUpSucceed, isSignUpError]);
+  const {
+    handleSignUp,
+    emailValidate,
+    handleEmailBlur,
+    nameValidate,
+    handleNameBlur,
+    passwordValidate,
+    handlePasswordBlur,
+    passwordCheckValidate,
+    handlePasswordCheckBlur,
+    passwordRef,
+  } = useSignUpPage();
 
   return (
     <Styled.Wrapper>
       <Title contents="회원가입" />
-      <Form onSubmit={handleSubmit}>
+      <Form onSubmit={handleSignUp}>
         <Fieldset>
           <Input
             description="이메일"
-            placeholder="coke@coke.com"
+            placeholder="example@example.com"
             onBlur={handleEmailBlur()}
             type="email"
             name="email"
@@ -102,7 +54,7 @@ const SignUp = () => {
         </Fieldset>
         <Fieldset>
           <Input
-            ref={pwd}
+            ref={passwordRef}
             description="비밀번호"
             placeholder="비밀번호를 입력해주세요."
             onBlur={handlePasswordBlur()}
@@ -118,7 +70,7 @@ const SignUp = () => {
           <Input
             description="비밀번호 확인"
             placeholder="비밀번호를 확인해주세요."
-            onBlur={handlePasswordCheckBlur(pwd.current?.value)}
+            onBlur={handlePasswordCheckBlur(passwordRef.current?.value)}
             name="passwordCheck"
             type="password"
           />
